+++ frontend/images/index.js
@@ -4,3 +4,4 @@ import "./icons/en.png";
 import "./icons/ru.png";
 import "./icons/google.svg";
 import "./icons/vk.svg";
+import "./icons/yandex.svg";